Better CI job names
Our CI configuration uses exact versions. This helps us with the
determinism of our CI, making it less dependent on external changes,
following the principle that CI should fail if and only if the code
inside a PR breaks something.

However, since exact versions are used to build job names, that means
everytime we bump the tested versions, the job names change and that
means we have to change our protected branch configuration to enforce
the checks with the new names.

With these changes I expect to keep the best of both worlds: CI
determinism, and stable and descriptive job names.